portfolioAPI/requirements.txt
python-dotenv
Django==4.2.9, <4.3
djangorestframework==3.14.0, <3.15
pillow==10.2.0, <10.3